Revision of sub-scope without CONTAINS block, inline example
Sub-scopes of a main scope without CONTAINS part were treated by creating
a virtual node containing part of the elements of the main scope. A major
drawback of this approach is that deleting an element from this virtual
node is not reflected in the complete node. A new approach has therefore
been developed, which uses a kind of view on the complete node that exposes
only part of its elements.